[routes]organise ProjectPostReply routes folder
