+++ backend/src/app.ts
@@ -6,11 +6,11 @@ import dashboardRouter from "./routes/index";
 import usersRouter from "./routes/Users.routes";
 import projectsRouter from "./routes/Projects.routes";
 import projectPostRouter from "./routes/ProjectPosts.routes";
-import projectPostReplyRouter from "./routes/ProjectPostReply.routes";
+import projectPostReplyRouter from "./routes/ProjectPostReply";
 import githubRouter from "./routes/Auth.routes";
 import friendshipsRouter from "./routes/Friendship";
 import userPostsRouter from "./routes/UserPosts.routes";
-import messagesRouter from "./routes/Messages.routes";
+import messagesRouter from "./routes/Messages";
 import fileRouter from "./routes/FileUploader";
 import searchRouter from "./routes/Search.routes";
 import matchRouter from "./routes/Match";
